Rewrite most of the game code
I wanted to revive this project as an easy way to refresh myself on
some pygame basics.

The rewrite improves the organization from the last rewrite and also
bring in most of the features from the original version of the game I
wrote.  The major pieces missing are sound effects and automated
paddle controls.

Collision is still wonky(wrong) on the paddle sides, and overly
simplified even on a normal hit.